Hecho cambios en varias de las clases

+++ ApiMegaTech/src/main/java/com/example/ApiMegaTech/Services/CategoryService.java
@@ -44,10 +44,9 @@ public class CategoryService {
     }
 
     public void deleteCategoryById(String id) {
-        try {
-            categoryRepository.deleteById(id);
-        }catch (Exception e){
-            System.out.println(e);
+        if (!categoryRepository.existsById(id)){
+            throw new RuntimeException();
         }
+        categoryRepository.deleteById(id);
     }
 }
